move crc64() to test fixture
