+++ test/fixtures/util.cpp
@@ -2,6 +2,11 @@
 
 #include <mbgl/platform/log.hpp>
 
+#pragma GCC diagnostic push
+#pragma GCC diagnostic ignored "-Wshadow"
+#include <boost/crc.hpp>
+#pragma GCC diagnostic pop
+
 #include <csignal>
 
 #include <unistd.h>
@@ -55,6 +60,17 @@ pid_t startServer(const char *executable) {
 
 void stopServer(pid_t pid) {
     kill(pid, SIGTERM);
+}
+
+// from https://gist.github.com/ArtemGr/997887
+uint64_t crc64(const char* data, size_t size) {
+    boost::crc_optimal<64, 0x04C11DB7, 0, 0, false, false> crc;
+    crc.process_bytes(data, size);
+    return crc.checksum();
+}
+
+uint64_t crc64(const std::string& str) {
+    return crc64(str.data(), str.size());
 }
 
 }
